requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
anthropic>=0.40.0
py-clob-client>=0.18.0
eth-account>=0.11.0
web3>=6.15.0
requests>=2.31.0
